Extract geometry functions out of interactive graph component file
